feature - Moodle 4.5.1+ with plugin updates.

+++ local/o365/export_manifest.php
@@ -27,7 +27,13 @@ require_once(__DIR__ . '/../../config.php');
 require_once($CFG->libdir . '/filestorage/zip_archive.php');
 require_once($CFG->dirroot . '/local/o365/lib.php');
 
+require_admin();
+
 // Mark manifest file as downloaded.
+$existingmanifestdownloadedsetting = get_config('local_o365', 'manifest_downloaded');
+if (!$existingmanifestdownloadedsetting) {
+    add_to_config_log('manifest_downloaded', $existingmanifestdownloadedsetting, true, 'local_o365');
+}
 set_config('manifest_downloaded', true, 'local_o365');
 purge_all_caches();
